Allow an explicit id value to be configured for the inline argument
Uses the same syntax as for passed in objects to the `files` argument. It should be noted that when passing in an invalid inline/files argument it no longer logs out to terminal but tries to read the argument, which will fail in most cases. This is more in line with PEP-20’s “Errors should never pass silently”.